Disable user supplied splicing sites on the opposite strand
In hybrid mode (-ub -uintrons.bed) user supplied splicing sites are added to list of canonical sites detected on one strand for the first alignement and then another alignement is made with canonical sites on the opposite strand. The best alignement is conserved.

When user sites for a transcript on one strand are added to the cannonical sites on the other strand this produces more splicing combinations and thus a better scoring but meaningless alignement.

This commit add logic discarding user provided sites that are canonical on the opposite strand from which the other cannonical sites are detected. This should help restore the balance and the detection of the correct transcript strand.